% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{butter}
\alias{butter}
\title{The butter function takes in a dataframe with 'words' and 'activation' columns, and
calls the spread function for a specific number of times to simulate the spread of
activation of a given initial activation space (specified in start_time) over time.}
\usage{
butter(start_run, retention = 0.5, network = gc.net, time = 10)
}
\arguments{
\item{start_run}{A non-empty dataframe with 'words' and 'activation' columns. Must be specified.}

\item{retention}{Proportion of activation that remains in the node, ranges from 0 to 1. Default is 0.5.}

\item{network}{Network where the spreading occurs. Must be specified.}

\item{time}{Number of time steps to run spread function for. Default is 10.}
}
\value{
A compiled dataframe with 'words', 'activation' and 'time' columns showing the spread of activation in the network over time.
}
\description{
The butter function takes in a dataframe with 'words' and 'activation' columns, and
calls the spread function for a specific number of times to simulate the spread of
activation of a given initial activation space (specified in start_time) over time.
}
\examples{
See Vignette for examples.
}
