% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.time.R
\name{estimate.time}
\alias{estimate.time}
\title{The estimate.time() function figures out, given a specified decay rate, the number of time steps
required for total activation in the network to decrease to a specified proportion of its value at
t = 0.}
\usage{
estimate.time(decay, final_proportion)
}
\arguments{
\item{decay}{Proportion of activation that is lost at each time step, ranges from 0 to 1.}

\item{final_proportion}{Proportion of initial total activation at t = 0 that is left at the end of the simulation}
}
\value{
An integer, # of timesteps to be used in simulation
}
\description{
The estimate.time() function figures out, given a specified decay rate, the number of time steps
required for total activation in the network to decrease to a specified proportion of its value at
t = 0.
}
\examples{
# estimate.time(0.2, 0.1) # 10
}
