% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.initial.R
\name{make.initial}
\alias{make.initial}
\title{The make.initial() function generates the initial activation space for a set of words based on
the target word. The basic idea is that when a target word is presented, not only the target word but
other words in the lexicon are also activated, albeit partially. The amount of partial activation
received by non-target words is inversely proportional to the edit distance between the nontarget
and specified target word. I.e., words that are immediate neighbors will receive 1/2 of the activation
received by the target, and words that have an edit distance of 4 will receive 1/5 of the activation
received by the target. In this function one must specify the number of activation units that the
target receives, and the nontarget words' initial activation will be calculated based on that value.
Another argument that could be interesting is to designate a cutoff for the number of nontarget words
to be (partially) activated, for e.g., neighbors only (cutoff = 2) or up to 3-hop neighbors (cutoff = 4).
Words with edit distance larger than or equal to the cutoff will have an initial activation value of 0.}
\usage{
make.initial(target, reference, target_act, cutoff = 100)
}
\arguments{
\item{target}{Character string of target node.}

\item{reference}{Reference lexicon to compare target node against.}

\item{target_act}{Number of activation units for the target word}

\item{cutoff}{Words with edit distance equal or larger than cutoff will not have an initial activation. To include all words, use a ridiculously large number for the cutoff, leave blank and use default of 100.}
}
\value{
A dataframe with 2 colummns words and activation. Be sure to convert the words column to the label column in hml before using it as the start_run dataframe in butter().
}
\description{
The make.initial() function generates the initial activation space for a set of words based on
the target word. The basic idea is that when a target word is presented, not only the target word but
other words in the lexicon are also activated, albeit partially. The amount of partial activation
received by non-target words is inversely proportional to the edit distance between the nontarget
and specified target word. I.e., words that are immediate neighbors will receive 1/2 of the activation
received by the target, and words that have an edit distance of 4 will receive 1/5 of the activation
received by the target. In this function one must specify the number of activation units that the
target receives, and the nontarget words' initial activation will be calculated based on that value.
Another argument that could be interesting is to designate a cutoff for the number of nontarget words
to be (partially) activated, for e.g., neighbors only (cutoff = 2) or up to 3-hop neighbors (cutoff = 4).
Words with edit distance larger than or equal to the cutoff will have an initial activation value of 0.
}
\examples{
# aback_initial <- make.initial('aback', gc$Ortho, 100, 5)
}
