\name{runSAM}
\alias{runSAM}
\title{Run the sam webapp}
\description{
Runs the sam web application for a graphical user interface.
}

\usage{
runSAM()
}

\details{Uses shiny to create a graphical user interface for SAM}

\references{Tusher, V.,  Tibshirani, R.  and Chu, G. (2001):
  Significance analysis of microarrays applied to the ionizing radiation response"  PNAS 2001 98: 5116-5121, (Apr 24).
  http://www-stat.stanford.edu/~tibs/sam}
\author{Michael J. Seo}


\examples{
\dontrun{runSAM()}
}
