% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{The spread function takes in a dataframe with 'words' and 'activation' columns, and
simulates the spread of activation of the words listed in the 'words' columns to its
neighbors in the specified network.}
\usage{
spread(start_run, retention, network)
}
\arguments{
\item{start_run}{A non-empty dataframe with 'words' and 'activation' columns.}

\item{retention}{Proportion of activation that remains in the node, ranges from 0 to 1.}

\item{network}{Network where the spreading occurs}
}
\value{
A updated dataframe with 'words' and 'activation' columns with new activation values.
}
\description{
The spread function takes in a dataframe with 'words' and 'activation' columns, and
simulates the spread of activation of the words listed in the 'words' columns to its
neighbors in the specified network.
}
\examples{
See Vignette for examples.
}
